Copy only the commands, and give each OS its own hero commands
Three defects from an independent acceptance pass over the deployed site.

1. "Copy command" copied the sample output too.

   The codeblock branch kept every non-empty line, stripped `$`, and joined
   the lot. On the CLI example that produced the two commands *plus* the
   JSON they print, so pasting into a shell ran the scan and then tried to
   execute `[`, `{ "port": 80, … }`, `]` and three hostnames.

   A block that uses `$` prompts now yields only those lines; a block with
   no prompts (the MCP config JSON) is copied whole, minus `//` comment
   lines. Preserving those lines verbatim also keeps the JSON's indentation,
   which the old per-line `.trim()` flattened.

   Verified by wrapping `navigator.clipboard.writeText` on the deployed page
   and capturing the real payload:
     CLI block  -> "netscli scan demo.local -p 80,443 --json\nnetscli
                    discover --json | jq '.[].hostname'"
     MCP block  -> the full JSON, indentation intact

2. Windows visitors were led with a bash pipeline.

   The first and most prominent hero box always held `hero.quickInstall`
   (`curl … install.sh | bash`), whatever the platform. On Windows that is
   the one command on the page that cannot run, with the correct `winget`
   line demoted to the smaller box below.

3. Linux visitors saw the same command twice.

   `packageCommands.linux` was byte-identical to `hero.quickInstall`, so
   both hero boxes printed the same `curl … | bash`.

   Both boxes are now OS-aware and always differ: primary is whatever the
   install section recommends for that platform, so hero and "Get started"
   agree, and secondary is a genuinely different route.

     windows  winget install fstubner.netscli   /  iwr … install.ps1 | iex
     macos    brew tap … && brew install …      /  curl … install.sh | bash
     linux    curl … install.sh | bash          /  yay -S netscli-bin

   Verified on the deployed build: Windows UA gives winget + iwr, Android
   (which now resolves to Linux) gives curl + yay, neither duplicated.

Copy-button setup moved to landing/copy-buttons.ts -- landing-page.ts
crossed the 300-line guard, and this is a self-contained concern that owns
the button markup, its states, and the what-to-copy decision. Re-verified
after the split that all 37 buttons still inject, both payloads are still
right, and the lightbox still opens and closes.

astro check clean, build passes, a11y 14 routes x both themes 0 violations,
file-size guard clean.
